moved semantic search functions to utils.semantic_search
